feat. conectando tabela de pedidos aos dados do backend através do service
